// app/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export interface Database {
// ...
  public: {
    Tables: {
// ...
      toys: {
        Row: {
          created_at: string;
          description: string | null;
          id: string;
          images: Json | null;
          is_rented: boolean;
          price: number;
          product_name: string;
        };
        Insert: {
          created_at?: string;
          description?: string | null;
          id: string;
          images?: Json | null;
          is_rented?: boolean;
          price: number;
          product_name: string;
        };
        Update: {
          created_at?: string;
          description?: string | null;
          id?: string;
          images?: Json | null;
          is_rented?: boolean;
          price?: number;
          product_name?: string;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
// ...
}